Added ST-XX
It adds a whopping near 86 MB (in full) but it's worth it.

Don't worry, this wasn't all manually typed! I made a (currently private) GDScript to make the process as easy and automatic as possible.

Only five packs are included in the repo since doing them all would be too much. You can get the rest from the link below, just make sure to rename the folders to be lowercase!
https://archive.org/details/AmigaSTXX